feat: Implement core database structure and seeders
- Add migrations for role, menu, role_menu, and user tables
- Create seeders with initial data for roles and menus
- Add database initialization scripts and VS Code tasks
- Fix database configuration for proper connectivity
- Set up initial admin user with default credentials

Migration timestamps set to 2025 for better version control

+++ app/Config/Database.php
@@ -47,7 +47,7 @@ class Database extends Config
             'compress'     => false,
             'strictOn'     => false,
             'failover'     => [],
-            'port'         => env('DB_PORT', 3306),
+            'port'         => (int) env('DB_PORT', 3306),
             'numberNative' => false,
             'dateFormat'   => [
                 'date'     => 'Y-m-d',
